requirements.txt
flask
flask-session
flask-cors
firebase-admin
numpy
polars
fuzzywuzzy
tqdm
gunicorn
google-cloud-logging